requirements.txt
Flask==3.0.0
flask-cors==4.0.0

# Firebase (Auth, Firestore, Storage)
firebase-admin==6.4.0

# Environment variables
python-dotenv==1.0.1

# PDF processing
PyMuPDF==1.24.1

numpy==1.26.4

# Vector database
chromadb==0.4.24

# AI / LLM
google-generativeai==0.8.3

# Utilities
requests==2.32.3

